Dynamically adjust table column width based on widget width
The first column of the process tables (CPU, memory, storage) shrinks or exapnds based on the window width.
The remaining columns are set to a fixed width. Also actually center the middle labels above memory and storage
bar graphs, so the middle labels don't shift left when the right labels take up more space as they change.
Minor tweaks to the formatting.